+++ gpaw/gpu/cpupy/__init__.py
@@ -38,10 +38,13 @@ def asnumpy(a, out=None):
     return out
 
 
-def asarray(a):
+def asarray(a, dtype=None):
     if isinstance(a, ndarray):
-        return a
-    return ndarray(np.array(a))
+        if a.dtype == dtype or dtype is None:
+            return a
+        else:
+            return ndarray(a._data.astype(dtype))
+    return ndarray(np.array(a, dtype=dtype))
 
 
 def array(a, dtype=None):
@@ -58,6 +61,10 @@ def ascontiguousarray(a):
 
 def dot(a, b):
     return ndarray(np.dot(a._data, b._data))
+
+
+def inner(a, b):
+    return ndarray(np.inner(a._data, b._data))
 
 
 def outer(a, b):
@@ -89,6 +96,14 @@ def abs(a):
 
 def exp(a):
     return ndarray(np.exp(a._data))
+
+
+def conjugate(a):
+    return ndarray(np.conjugate(a._data))
+
+
+def log(a):
+    return ndarray(np.log(a._data))
 
 
 def eye(n):
@@ -123,7 +138,8 @@ def fuse():
 class ndarray:
     def __init__(self, data):
         if isinstance(data, (float, complex, int, np.int32, np.int64,
-                             np.bool_)):
+                             np.bool_, np.float64, np.float32,
+                             np.complex64, np.complex128)):
             data = np.asarray(data)
         assert isinstance(data, np.ndarray), type(data)
         self._data = data
@@ -168,6 +184,9 @@ class ndarray:
 
     def copy(self):
         return ndarray(self._data.copy())
+
+    def astype(self, dtype):
+        return ndarray(self._data.astype(dtype))
 
     def all(self):
         return ndarray(self._data.all())
@@ -230,6 +249,26 @@ class ndarray:
         if isinstance(other, (float, complex, int)):
             return self._data != other
         return ndarray(self._data != other._data)
+
+    def __lt__(self, other):
+        if isinstance(other, (float, complex, int)):
+            return self._data < other
+        return ndarray(self._data < other._data)
+
+    def __le__(self, other):
+        if isinstance(other, (float, complex, int)):
+            return self._data <= other
+        return ndarray(self._data <= other._data)
+
+    def __gt__(self, other):
+        if isinstance(other, (float, complex, int)):
+            return self._data > other
+        return ndarray(self._data > other._data)
+
+    def __ge__(self, other):
+        if isinstance(other, (float, complex, int)):
+            return self._data >= other
+        return ndarray(self._data >= other._data)
 
     def __neg__(self):
         return ndarray(-self._data)