+++ doc/releasenotes.rst
@@ -10,7 +10,11 @@ Git master branch
 
 :git:`master <>`.
 
+* The fluctuation-dissipation theorem has been added to the SJM.
+
 * Minimum version requirements: Python 3.9, ASE 3.23.0.
+
+* The constant inner potential DFT method has been included SJM.
 
 * Sign of electron chemical potential in SJM text output has been corrected.
   The resulting (Grand) energies are not affected by this change.
@@ -24,6 +28,7 @@ Git master branch
   **k**-points has been implemented.  See :ref:`hse06 on lda` and
   :class:`gpaw.new.pw.nschse.NonSelfConsistentHSE06`.
 
+* Bug fix for spin-polarized LCAO-TDDFT circular dichroism See :mr:`2667`.
 
 Version 25.1.0
 ==============